Don't try to modify undef.
Case 100217: When an argument is passed to a subroutine, it is passed by
reference.  Perl before 5.19 contained a bug where taking a reference to
undef would return a new scalar.  This bug has been fixed.

IPC::Pipeline accepted undef as an argument but then tried to modify it.
Since undef is read-only, this attempt failed.  Wrap the assignment in an
eval to avoid throwing a fatal error.

Also, bump the version in preparation for a release.

This patch is based off of one kindly provided by Father Chrysostomos.